new feature: rent info -> bill

+++ BE/httc-sport/src/main/java/com/fpoly/httc_sport/entity/Bill.java
@@ -38,7 +38,7 @@ public class Bill implements Serializable {
 	LocalTime startTime;
 	@Temporal(TemporalType.TIME)
 	LocalTime endTime;
-	Double total;
+	Integer total;
 	@Builder.Default
 	Boolean isRate = false;
